Update for NL000093 and NL000319 Fix
Included suggestions by @ptatian. One issue is that the following
warning:

"WARNING: The data set PRESCAT.BUILDING_GEOCODE is not sorted by the ID
variables. Observations will be matched using NOTSORTED logic with the
assumption that the observations correspond one-to-one."

This occurs when the altered building_geocode file is compared to the
prescat.building_geocode file. However, my understanding is that the
prescat file is already sorted by nlihc_id bldg_addre.